fix(pets): use sizeof(float) in SetActions memcpy for m_speeds
Copy-paste from the m_actions line left sizeof(int) instead of
sizeof(float). Both are 4 bytes on all targets so no runtime effect,
but semantically wrong and misleading when reading the code.

+++ src/source/Gameplay/Pets/w_PetProcess.cpp
@@ -45,7 +45,7 @@ void PetInfo::SetActions(int count, int* actions, float* speeds)
     memcpy(m_actions, actions, sizeof(int) * m_count);
 
     m_speeds = new float[count];
-    memcpy(m_speeds, speeds, sizeof(int) * m_count);
+    memcpy(m_speeds, speeds, sizeof(float) * m_count);
 }
 
 PetProcessPtr g_petProcess;
